+++ lib/Circle.js
@@ -1,0 +1,8 @@
+class Circle {
+    constructor(shape,color) {
+        this.shape=shape;
+        this.color=color;
+    }
+}
+
+module.exports = Circle;
