+++ lib/Square.js
@@ -1,0 +1,10 @@
+class Square {
+    constructor(shape,color) {
+        this.shape=shape;
+        this.color=color;
+    }
+}
+
+
+
+module.exports = Square;
