+++ lib/Triangle.js
@@ -1,0 +1,8 @@
+class Triangle {
+    constructor(shape,color) {
+        this.shape=shape;
+        this.color=color;
+    }
+}
+
+module.exports = Triangle;
