--- lib/shape.js
@@ -1,20 +1,0 @@
-class Square {
-    constructor(shape,color) {
-        this.shape=shape;
-        this.color=color;
-    }
-}
-
-class Triangle {
-    constructor(shape,color) {
-        this.shape=shape;
-        this.color=color;
-    }
-}
-
-class Circle {
-    constructor(shape,color) {
-        this.shape=shape;
-        this.color=color;
-    }
-}
